remote-playback/event-handlers-manual.html

<!DOCTYPE html>
<html>
  <title>
    Test that all event handlers are called.
  </title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/common/media.js"></script>
  <body>
    <div id="prep">
      <p>Please make sure a device for remote playback is <b>available.</b></p>
      <button id="prompt-button-prep">Show devices</button>
      <button id="start-button">Start test</button>
    </div>
    <div id="pick-device" style="display: none">
      <p>
        Click the button below to prompt for a remote playback device and select
        one! After connecting to the device, please click again and disconnect
        from device.
      </p>
      <button id="prompt-button">Pick device</button>
      <button id="disconnect-button" style="display: none">
        Disconnect device
      </button>
    </div>
  </body>
  <script>
    var v = document.createElement("video");
    v.src = getVideoURI("/media/movie_5");

    var startButton = document.getElementById("start-button");
    startButton.onclick = function () {
      var prep = document.getElementById("prep");
      prep.style.display = "none";
      var pickDeviceOptions = document.getElementById("pick-device");
      pickDeviceOptions.style.display = "block";
    };
    var promptPrepButton = document.getElementById("prompt-button-prep");
    promptPrepButton.onclick = function () {
      v.remote
        .prompt()
        .then(() => {})
        .catch(() => {});
    };
    setup({
      explicit_timeout: true,
    });
    async_test((t) => {
      var onConnectingCalled = false;
      var onConnectCalled = false;
      var onDisconnectCalled = false;
      v.remote.onconnecting = function () {
        onConnectingCalled = true;
      };

      v.remote.onconnect = function () {
        onConnectCalled = true;
      };

      v.remote.ondisconnect = function () {
        onDisconnectCalled = true;
      };

      var button = document.getElementById("prompt-button");
      var disconnectButton = document.getElementById("disconnect-button");
      button.onclick = function () {
        promise_test(() => {
          return v.remote.prompt().then(() => {
            button.style.display = "none";
            disconnectButton.style.display = "inline";
          });
        }, "Prompt to connect device");
      };
      disconnectButton.onclick = function () {
        v.remote.prompt().then(t.step_func_done(() => {
          assert_true(onConnectingCalled, "onconnecting was called");
          assert_true(onConnectCalled, "onconnect was called");
          assert_true(onDisconnectCalled, "ondisconnect was called");
        }));
      };
    }, "Test that all event handlers are called.");
  </script>
</html>
